doub

+++ src/StockPurchase.java
@@ -2,10 +2,10 @@ import java.util.Date;
 
 /**
  * Class representing the purchase of a stock in a portfolio, it contains
- * the amount of shares purchaseDate and sellDate.
+ * the amount of shares purchaseDate.
  */
 public class StockPurchase {
-  int shares;
+  double shares;
   Date purchaseDate;
 
   /**
@@ -14,7 +14,7 @@ public class StockPurchase {
    * @param shares the amount of shares being purchased.
    * @param purchaseDate the date of the purchase.
    */
-  StockPurchase(int shares, Date purchaseDate) {
+  StockPurchase(double shares, Date purchaseDate) {
     this.shares = shares;
     this.purchaseDate = purchaseDate;
   }
